Start to change the way input works
Before you'd enter your cat in the input box, then when you were done
you'd somehow select it and it would be added to your selected cats.
This model, I think, is confusing. In this branch I'm trying to have a
'preview box' that mirrors what you're typing, but how it would appear
once you submit it. The suggestion box will still appear, but selected
suggestions will be added to your ongoing text input, instead of
clearing it and starting over.